refactor: moved handlers in routes, config file and startup file
