+++ Cargo.toml
@@ -16,5 +16,17 @@ actix-web = "4"
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 serde = { version = "1", features = ["derive"] }
 
+[dependencies.sqlx]
+version = "0.7"
+default-features = false
+features = [
+    "postgres", 
+    "runtime-tokio-rustls", 
+    "macros",
+    "uuid",
+    "chrono",
+    "migrate"
+]
+
 [dev-dependencies]
 reqwest = "0.11"
