+++ src/startup.rs
@@ -1,0 +1,18 @@
+use std::net::TcpListener;
+
+use actix_web::dev::Server;
+use actix_web::{web, App, HttpServer};
+
+use crate::routes::{health_check, subscribe};
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/subscribe", web::post().to(subscribe))
+    })
+    .listen(listener)?
+    .run();
+
+    Ok(server)
+}
